fix(same-file-same-folder): fix an issue with naming
* fix(files): didnt send all the files to npm

* fix(fix docs): fix documentation (lack of docs at all) and example

* fix(docs): errors in documentation didnt allow for nice reading

* docs(readme.md): i added a space to prove a point

Proving a point is fun. Is the lofical way of saying i told you so

* feat(underscore support): add underscore support on the file naming

add underscore support on file name. this might still be a bad thing if two files collide. for now
this will do but it might be better to, eventually add some check before-hand

* fix(same-file-same-folder): fix an issue with naming

fix an issue in which if a file has the same name as the folder, the file name would be the one that
would disappear

fix #6